PopUpContentAddFlight component created within new PopUpBox component
