+++ frontend-app-skynuc/src/components/popUpBox/popUpBox.css
